add face-to-face support section to triage tool
This commit adds a 'Plan' container react component for the detailed
plan section of the triage tool, along with an initial FaceToFaceSupport
component.

It also defines a planTemplates object in the triage tool config, this
is for simple string template functions that have different output based
on state in the mobx-state-tree store.